HTTPS Proxy: Fix large body request.
When request need to be stored in a temp file, the
ngx.req.get_body_data() will return nil. On some uses cases, users
report that large body is not correctly set. So things are failing on
their end.

This commit will check if the file has been created, read the data from
there and set the body correctly.

Fix THREESCALE-3863

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>