OIDC: fix issue on remote v2.
This is a second patch, and it's related to PR #1304.

When using remotev2 loader, the config.oidc is created in there, and
there is also a caching mechanism for OIDC[0], so we need to append the
service_id and send a valid table in case of non OIDC setup.

Because this oidc is cached, we also do a deepcopy, to avoid to use
always the same reference, so the config.oidc[i].service_id is always
different and we didn't get a invalid filtering.

This is only failing when using APICAST_SERVICE_LIST

[0] Commit: 720cd99f5dcbd2e11e3b3f07d5ee5888865740d7
Related to: THREESCALE-6042

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>