gateway/src/apicast/mapping_rule: fix error when passing nil args to qs matching function
When a call to match a rule is made without specifying the query string
args or setting them as nil, which is equivalent, _and_ the match is
successful for the method and the path/uri, then the function
`self.querystring_params` will be called with the nil args, which in
turn will call `matches_querystring_params` with a nil table which will
crash on trying to index with a parameter: `local value = args[param]`.

While I haven't come across a case in which this is used currently in
the code in this fashion, it is certainly used in this way in the specs,
relying on the fact that some code paths will return early with a non
match before evaluating the query string parameters.

Let's use an empty table to keep the current behavior rather than
crashing or returning true/false.